+++ src/codesyncd/populate_buffer.ts
@@ -67,7 +67,7 @@ export const populateBufferForMissedEvents = async (readyRepos: any) => {
                 if (timeTook > GLOB_TIME_TAKEN_THRESHOLD) {
                     CodeSyncLogger.warning(`populateBuffer took=${timeTook}s for ${repoPath}, files=${obj.itemPaths.length}, uuid=${instanceUUID}`);
                 }
-            }    
+            }
             const generateDiffForDeletedFilesKey = `${repoPath}:${branch}:generateDiffForDeletedFiles`;
             const canSkipDeleteHandler = CodeSyncState.canSkipRun(generateDiffForDeletedFilesKey, RUN_DELETE_HANDLER_AFTER);
             if (canSkipDeleteHandler) continue;
@@ -164,7 +164,7 @@ class PopulateBuffer {
 
     async populateBufferForRepo() {
         const instanceUUID = CodeSyncState.get(CODESYNC_STATES.INSTANCE_UUID);
-        CodeSyncLogger.debug(`Watching Repo: ${this.repoPath}, uuid=${instanceUUID}`);
+        CodeSyncLogger.debug(`Watching Repo: ${this.repoPath},  branch=${this.branch}, files=${this.itemPaths.length}, uuid=${instanceUUID}`);
         const handler = new eventHandler(this.repoPath, "", true);
         for (const itemPath of this.itemPaths) {
             let isRename = false;
@@ -251,9 +251,9 @@ class PopulateBuffer {
                 removeFile(shadowFilePath, "getPotentialRenamedFiles");
                 return false;
             }
-            const ignoreAblePath = shouldIgnorePath(relPath, this.defaultIgnorePatterns, this.syncIgnoreItems);
+            const ignorablePath = shouldIgnorePath(relPath, this.defaultIgnorePatterns, this.syncIgnoreItems);
             // If file is not in config OR is is present in .syncignore, remove the file from .shadow
-            if (ignoreAblePath) {
+            if (ignorablePath) {
                 removeFile(shadowFilePath, "getPotentialRenamedFiles");
                 return false;
             }
@@ -318,8 +318,8 @@ class PopulateBuffer {
             const cacheFilePath = path.join(this.deletedRepoBranchPath, relPath);
             const shadowFilePath = path.join(this.shadowRepoBranchPath, relPath);
             // See if should discard this file
-            const isIgnoreablePath = shouldIgnorePath(relPath, this.defaultIgnorePatterns, this.syncIgnoreItems);
-            if (isIgnoreablePath  ||
+            const isIgnorablePath = shouldIgnorePath(relPath, this.defaultIgnorePatterns, this.syncIgnoreItems);
+            if (isIgnorablePath  ||
                 activeRelPaths.includes(relPath) ||
                 this.renamedFiles.includes(relPath) ||
                 fs.existsSync(cacheFilePath) ||